ottos-expeditions/pyproject.toml
[project]
name = "ottos-expeditions"
version = "0.0.3"
authors = [{ name = "Ascend", email = "support@ascend.io" }]
description = "Otto's Expeditions"
readme = "README.md"
requires-python = ">=3.9"
classifiers = [
  "Programming Language :: Python :: 3",
  "License :: OSI Approved :: MIT License",
  "Operating System :: OS Independent",
]
dependencies = [
  # local secrets
  'python-dotenv',
  # project config
  'PyYAML',
  # CLI
  'rich',
  'typer',
  # GUI
  'streamlit',
  'shiny',
  'shinyswatch',
  'shinywidgets',
  # data
  'ibis-framework[sqlite,duckdb,datafusion,polars,bigquery,snowflake,deltalake]',
  'snowflake-snowpark-python',
  # synthetic data
  'Faker',
  # viz
  'plotly',
  # AI
  'pydantic-ai',
]

[project.optional-dependencies]
dev = ["ruff", "ipython", "jupyter", "nbclient", "ipykernel"]
test = ["pytest"] # for appearance only

[project.urls]
"Homepage" = "https://github.com/ascend-io/ascend-community"
"Bug Tracker" = "https://github.com/ascend-io/ascend-community/issues"

[project.scripts]
"ottos-expeditions" = "ottos_expeditions.ui.cli:app"

[build-system]
# TODO: update to uv build system once available
# https://github.com/astral-sh/uv/issues/3957
requires = ["hatchling"]
build-backend = "hatchling.build"
